Move clone functional tests outside of storage sub-package
