+++ tests/BUILD.bazel
@@ -108,6 +108,7 @@ go_test(
     srcs = [
         "access_test.go",
         "canary_upgrade_test.go",
+        "clone_test.go",
         "config_test.go",
         "console_test.go",
         "container_disk_test.go",
@@ -158,9 +159,9 @@ go_test(
         "vnc_test.go",
         "windows_test.go",
     ],
+    embed = [":go_default_library"],
     visibility = ["//visibility:public"],
     deps = [
-        ":go_default_library",
         "//pkg/certificates/triple:go_default_library",
         "//pkg/certificates/triple/cert:go_default_library",
         "//pkg/cloud-init:go_default_library",
@@ -195,6 +196,7 @@ go_test(
         "//pkg/virtctl/pause:go_default_library",
         "//pkg/virtctl/softreboot:go_default_library",
         "//pkg/virtctl/vm:go_default_library",
+        "//staging/src/kubevirt.io/api/clone/v1alpha1:go_default_library",
         "//staging/src/kubevirt.io/api/core:go_default_library",
         "//staging/src/kubevirt.io/api/core/v1:go_default_library",
         "//staging/src/kubevirt.io/api/instancetype:go_default_library",
@@ -243,6 +245,7 @@ go_test(
         "//vendor/github.com/onsi/ginkgo/v2:go_default_library",
         "//vendor/github.com/onsi/ginkgo/v2/reporters:go_default_library",
         "//vendor/github.com/onsi/gomega:go_default_library",
+        "//vendor/github.com/onsi/gomega/format:go_default_library",
         "//vendor/github.com/onsi/gomega/gmeasure:go_default_library",
         "//vendor/github.com/onsi/gomega/gstruct:go_default_library",
         "//vendor/github.com/onsi/gomega/types:go_default_library",
