+++ tests/clone_test.go
@@ -1,4 +1,4 @@
-package storage
+package tests
 
 import (
 	"context"
@@ -20,7 +20,6 @@ import (
 	"kubevirt.io/client-go/kubecli"
 
 	virtconfig "kubevirt.io/kubevirt/pkg/virt-config"
-	"kubevirt.io/kubevirt/tests"
 	"kubevirt.io/kubevirt/tests/console"
 	cd "kubevirt.io/kubevirt/tests/containerdisk"
 	. "kubevirt.io/kubevirt/tests/framework/matcher"
@@ -35,7 +34,7 @@ const (
 
 type loginFunction func(*virtv1.VirtualMachineInstance) error
 
-var _ = SIGDescribe("[Serial]VirtualMachineClone Tests", func() {
+var _ = Describe("[Serial][sig-compute]VirtualMachineClone Tests", func() {
 	var err error
 	var virtClient kubecli.KubevirtClient
 
@@ -43,7 +42,7 @@ var _ = SIGDescribe("[Serial]VirtualMachineClone Tests", func() {
 		virtClient, err = kubecli.GetKubevirtClient()
 		util.PanicOnError(err)
 
-		tests.EnableFeatureGate(virtconfig.SnapshotGate)
+		EnableFeatureGate(virtconfig.SnapshotGate)
 
 		format.MaxLength = 0
 	})
@@ -51,7 +50,7 @@ var _ = SIGDescribe("[Serial]VirtualMachineClone Tests", func() {
 	createVM := func(options ...libvmi.Option) (vm *virtv1.VirtualMachine) {
 		vmi := libvmi.NewCirros(options...)
 		vmi.Namespace = util.NamespaceTestDefault
-		vm = tests.NewRandomVirtualMachine(vmi, false)
+		vm = NewRandomVirtualMachine(vmi, false)
 		vm.Annotations = vmi.Annotations
 		vm.Labels = vmi.Labels
 
@@ -96,14 +95,14 @@ var _ = SIGDescribe("[Serial]VirtualMachineClone Tests", func() {
 
 	expectVMRunnable := func(vm *virtv1.VirtualMachine, login loginFunction) *virtv1.VirtualMachine {
 		By(fmt.Sprintf("Starting VM %s", vm.Name))
-		vm = tests.StartVirtualMachine(vm)
+		vm = StartVirtualMachine(vm)
 		targetVMI, err := virtClient.VirtualMachineInstance(vm.Namespace).Get(vm.Name, &v1.GetOptions{})
 		Expect(err).ShouldNot(HaveOccurred())
 
 		err = login(targetVMI)
 		Expect(err).ShouldNot(HaveOccurred())
 
-		vm = tests.StopVirtualMachine(vm)
+		vm = StopVirtualMachine(vm)
 
 		return vm
 	}
@@ -349,7 +348,7 @@ var _ = SIGDescribe("[Serial]VirtualMachineClone Tests", func() {
 			}
 
 			createVMWithStorageClass := func(storageClass string, running bool) *virtv1.VirtualMachine {
-				vm := tests.NewRandomVMWithDataVolumeWithRegistryImport(
+				vm := NewRandomVMWithDataVolumeWithRegistryImport(
 					cd.DataVolumeImportUrlForContainerDisk(cd.ContainerDiskAlpine),
 					util.NamespaceTestDefault,
 					storageClass,
@@ -407,7 +406,7 @@ var _ = SIGDescribe("[Serial]VirtualMachineClone Tests", func() {
 			}
 
 			It("with a simple clone", func() {
-				snapshotStorageClass, err := getSnapshotStorageClass(virtClient)
+				snapshotStorageClass, err := libstorage.GetSnapshotStorageClass(virtClient)
 				Expect(err).ToNot(HaveOccurred())
 
 				if snapshotStorageClass == "" {
@@ -440,7 +439,7 @@ var _ = SIGDescribe("[Serial]VirtualMachineClone Tests", func() {
 				sourceVM = createVMWithStorageClass(sc, true)
 				sourceVM, err = virtClient.VirtualMachine(sourceVM.Namespace).Get(sourceVM.Name, &v1.GetOptions{})
 				Expect(err).ToNot(HaveOccurred())
-				sourceVM = tests.StopVirtualMachine(sourceVM)
+				sourceVM = StopVirtualMachine(sourceVM)
 
 				vmClone = generateClone()
 				vmClone, err = virtClient.VirtualMachineClone(vmClone.Namespace).Create(context.Background(), vmClone, v1.CreateOptions{})
