+++ tests/storage/BUILD.bazel
@@ -3,7 +3,6 @@ load("@io_bazel_rules_go//go:def.bzl", "go_library")
 go_library(
     name = "go_default_library",
     srcs = [
-        "clone.go",
         "datavolume.go",
         "events.go",
         "export.go",
